// src/api/generated/model/outreachGenerateIn.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */
import type { OutreachChannel } from './outreachChannel';
import type { OutreachTone } from './outreachTone';

export interface OutreachGenerateIn {
  service_id?: string | null;
  channel?: OutreachChannel;
  language?: string;
  tone?: OutreachTone;
  sender_name?: string | null;
  sender_title?: string | null;
  sender_company?: string;
}